Make fat_test app work again
I think the storage_master/fat_test app is the only one in all of Tock
that uses a folder inside the app to hold some source files. To handle
this, it needs some special commands in its local Makefile to create the
proper directories in the `build` folder.

However, we changed the TOCK_ARCHES stuff pretty heavily a while back,
which broke this hard. The best I could come up with was to replicate
the functionality in this Makefile, which is pretty heavy duty for
something as dumb as creating directories when needed.

Additionally, (and I'm hoping @ppannuto might have an idea here) the
original rule, which presumably worked at the time, matches against the
`%` wildcard. When run now, this creates a folder for each FILE.d
generated, which really bugs the build system because those are supposed
to be files... For now, I just hardcoded in rules for each folder we'd
want to make, but a general solution would be good. (and possibly maybe
allow us to solve the above problem too?)